requirements.txt
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
isort==5.12.0
mypy==1.5.1
httpx==0.25.2
pydantic==2.5.2
openai==1.76.0
python-dotenv==1.0.1
sentence-transformers==2.2.2
pinecone-client==3.0.0
